Add a function to classify identifiers.
This commit adds a function that takes a general identifier, determines
what kind of an identifier it is (e.g., DOI, ISBN, etc.), and then
normalizes the identifier.

This also fixes the `add` function to call the renamed `is_arxiv`
function.

Finally, this imports the various submodules in `zoia.ids` so that one
only needs to do:

```py
import zoia.ids
```

rather than needing to import each kind of ID individually.