Support adding items by ISBN.
This commit extends `zoia add` to support adding ISBNs.

This commit also adds `isbnlib` as a dependency and switches the ISBN
verification code to use this library.  The previous code did not
correctly handle 10 digit ISBNs since it did not perform the correct
conversion to ISBN-13 (and also would fail if there was an 'X' present
in the checksum digit).