Add a method to print metadata nicely.
This commit adds a `__str__` method to the `Metadatum` dataclass to
nicely print the entry in the format 'Author (YYYY), "TITLE"'.

The string may be abbreviated to keep from being too long, but at least
three words of the title are always retained.

Furthermore, when an item is successfully added, this string is now used
when printing the success message.