Add a YAML header section to new notes.
This commit adds a header section with some YAML-formatted metadata to
new notes.  Specifically, the following data are added:

* The title.
* The authors with up to four listed.  For papers with five or more
  authors, only the first three are listed, along with an "et al.".
* The year
* A list of tags (separated by commas).  To keep this from occupying too
  much vertical space these will appear on one line rather than a list.

Note that at the moment if this data is edited, those changes will not
be reflected in the metadata file.  Furthermore, the header is generated
only once, when a new note is created.  This can lead to inconsistencies
between the header and the metadata if the user edits the header, but
not the metadata.  In the future `zoia note` should regenerate this
header every time a note is opened, and write to the metadata when the
header section has changed.

This commit also adds a `zoia.yaml` utility to provide some default
formatting for YAML operations.  In particular, it does the following:

* Sets the default indentation to 4.
* Indents lists.