Regenerate note header on every `zoia note` call.
Prior to this commit the note header was only generated a single time
when the file was generated.  This changes the header to be regenerated
every time `zoia note` is called so that it is guaranteed to be up to
date.

A consequence of this is that the header is no longer stored in the file
itself.  Instead the `notes.md` file only stores the body of the note.

In order to accomplish this, this commit also adds a `remove_header`
function in `zoia.yaml`.

A second consequence of these changes is that the user no longer edits
the notes file directly.  Instead they edit a temporary file, the header
is discarded, and the body is written to `notes.md`.  At the moment any
changes to the header metadata are ignored.  A future commit will likely
update metadata based on edits to the note header.