Extend `zoia add` to handle PDFs.
In order to accomplish this, this commit also changes the various
downloaders to also save an MD5 hash of downloaded PDFs.  This allows
zoia to quickly check whether or not the given PDF already exisis in the
library.

This commit also adds a utility called `zoia.yaml.edit_until_valid`
which will prompt the user to edit some YAML text, but will give them
the chance to continue editing if there is a YAML syntax error.

A few things are currently missing:

* Tests of the `zoia.add._add_pdf` functionality need to be added.
* If the PDF already exists, the error message needs to be improved to
  specify which item this is.
* It should be possible to add both a DOI and a PDF at the same time.