Confirm a paper's metadata when adding from PDF.
This commit improves the way Zoia adds PDFs by explictly stating what
Zoia believes the paper to be when it has found a possible DOI and
asking for confirmation from the user.

If the user does not believe the metadata are correct, Zoia now gives
the user the option of filling out the metadata manually.

This commit also adds a unit test for the `_add_pdf` function.