Update the metadata API.
This commit makes a few changes to the `zoia.backend.metadata` API.
Most notably it renames `load_metadata` to `get_all_metadata`.

This commit also adds a function `get_metadata` which will get the
metadata for a specific citekey.  Prior to this commit, the only way of
doing this was to call `load_metadata` and then query the particular
citekey.

Since `zoia`'s backend is going to move to Redis shortly, this
change will minimize any API changes needed for this transition.

This commit also adds a function `commit_exists` that checks whether or
not a citekey is present in the library.