Rearchitect the metadata.
This commit applies a major rearchitecture to the metadata API.  In
preparation for using a more sophisticated backend this switches from a
functional API to an object-oriented API.  The main motivation here is
that it makes initialization of the service more explicit.

Now there is an abstract `Metadata` class that various backends can
inherit from.  Currently only the `JSONMetadata` backend is implemented.

A number of files needed to be touched to implement these changes.  The
functionality of the CLI hasn't changed.

One other change is that the metadata file is now stored in a separate
location from the library.  Note that this is a breaking change.
Libraries created prior to this commit will not work afterwards.  In
order to migrate you need to move the `.metadata.json` file to
`$HOME/.local/share/zoia/`.