Add a SQLite backend.
This commit adds a backend that uses a SQLite database.

In order to get this to work efficiently one small change needed to be
made to the `Metadata` API.  Instead of providing a method like
`arxiv_ids` that returns a set of all existing arXiv IDs, the `Metadata`
abstract class now has a `arxiv_id_exists` method that checks the
existence of a provided ID.  These methods were only used for existence
checks so only minor tweaks needed to be made to the method calls to
incorporate this change.

Note that there are currently no unit tests for this new backend.  I
have tried running it on a single article to make sure that it runs, but
there may still be bugs.