Add a SQLite backend.
This commit adds a backend that uses a SQLite database.

In order to get this to work efficiently one small change needed to be
made to the `Metadata` API.  Instead of providing a method like
`arxiv_ids` that returns a set of all existing arXiv IDs, the `Metadata`
abstract class now has a `arxiv_id_exists` method that checks the
existence of a provided ID.  These methods were only used for existence
checks so only minor tweaks needed to be made to the method calls to
incorporate this change.

Otherwise the metadata API is unchanged and so no other downstream
changes would need to be made with this backend switch.

By default all new `zoia` libraries will use this new SQLite backend.

This commit adds a suite of unit tests for this backend.