Add a __setitem__ method to `Metadata` class.
This commit adds a `__setitem__` method to the various metadata classes.
This new method replaces the previous `append` and `replace` methods.

This method is "safe" in the sense that if the caller provides a
dictionary that does not change certain keys it will not remove those
keys.  The `__setitem__` method should therefore be thought of as more
of an update method rather than an assignment method.