Refactor `zoia add`.
This commit moves the work to add a new entry into `zoia.backend`.  The
intent here is to make it easier to perform the same task across
different user interfaces (the CLI, the TUI, and the GUI).  Though
`zoia` only supports a CLI at the moment, this commit is a step towards
providing a TUI and GUI.

In order to make this possible this commit makes a more complete
separation of the backend work vs. displaying status messages to the
user.  In order to do this this commit adds a `StatusMessage` class that
is intended to be a common interface to display status messages.  At the
moment this only uses the `Halo` spinner, but in the future this will
display spinners in the TUI & GUI.

Note that this refactor is not 100% complete.  At the moment if the user
adds a PDF and `zoia` finds a DOI, the backend will still ask for
confirmation that the DOI is correct via `click`.