Remove the space before the ellipses.
When an entry is printed the title may be abbreviated if the string is
too long.  This abbrevation is done by printing ellipses for the rest of
the title.  Prior to this commit, however, there was a space between the
final word and the ellipses.  This commit eliminates this space.

Before: "On the electrodynamics ..."

After: "On the electrodynamics..."